fix(Equipments): get last update date
