+++ src/v2/equipments/routes/equipments.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import { adaptRouteV2 } from "../../../server/http/adapters/express-route.adapter";
 import { EquipmentsControllers } from "../controllers";
+import { authorization } from "../../../server/http/http-middlewares";
 
 export const setupEquipmentsRoutes = (router: Router): void => {
     router.get(
@@ -17,9 +18,14 @@ export const setupEquipmentsRoutes = (router: Router): void => {
     );
 
     router.get(
+        "/last-update-date",
+        authorization,
+        adaptRouteV2(EquipmentsControllers.getDateOfLastMeasurementTaken)
+    );
+
+    router.get(
         "/meteorological_organ/access_credentials",
         adaptRouteV2(EquipmentsControllers.getMeteorologicalOrganAccessCredentials)
     );
-
 
 };
